refactor(oas): split monolithic `DefaultConverter`
relates-to #120
