+++ packages/oas/src/converter/subconverters/SubConverter.ts
@@ -1,0 +1,3 @@
+export interface SubConverter<R> {
+  convert(path: string, method: string): R;
+}
