+++ packages/oas/src/converter/subconverters/SubPart.ts
@@ -1,0 +1,6 @@
+export enum SubPart {
+  HEADERS = 'HEADERS',
+  PATH = 'PATH',
+  POST_DATA = 'POST_DATA',
+  QUERY_STRING = 'QUERY_STRING'
+}
